Pass whole extensions rather than just files
This ensures that all concents of enablement are always handled once and
by the ExtensionEnablementManager (removing the old annotateActiveExtensions).
It also paves the way for just-in-time evaluation of extension context
file inclusion of active extensions rather than having to replumb all of
the files through activation scenarios.

For tools added via extension, only discover them if they are marked as
from an active extension and dynamically fetch the list of extensions
from the config object when invoking the discovery method. This should
allow us to manipulate the extension collection on config and accurately
refresh the list of available tools.